Vitaly/embedded camera (#115)
* Fixed minor compilation error for ARM

* WIP: embedded camera demo

* Added factories for creating angular position input and output modules
Added command-line options parsing

* Embedded camera demo: removed unused class member

* Camera demo: added script for updating udev rules

* Camera demo: added platform-agnostic startup script

* Camera demo: WIP added frame rate parameter. Note that tested webcams do not allow user-set frame rates.

* Camera demo: enabled Pi Servo in the position output creator

* Embedded camera demo: added readme

* Embedded camera demo: added verbose flag to the orientation output

+++ demos/channel_demo/src/reader.c
@@ -167,10 +167,10 @@ static int reader_run_perf_test(reader_t *reader) {
         ++reader->counts[*msg_idx];
     }
 
-    snprintf(prefix, sizeof(prefix) - 1, "perf_%u_%u_%zu", 
+    snprintf(prefix, sizeof(prefix) - 1, "perf_%u_%u_%llu", 
         reader->common.data.perf.len,
         reader->common.data.perf.count,
-        reader->common.data.delay_ns);
+        (unsigned long long)reader->common.data.delay_ns);
 
     if (bin_save(prefix, reader->common.data.out_dir, reader->counts, 
                     reader->common.data.perf.count) != 0) {
